+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.1.2] - 2023-07-27
+
 ### Fixed
 
 - Enabled direct access to fields inside a FieldGroupContent model using a field name
@@ -172,7 +174,9 @@
 
 - Initial Beta Release
 
-[Unreleased]: https://github.com/ExpressionEngine/Coilpack/compare/1.1.1...HEAD
+[Unreleased]: https://github.com/ExpressionEngine/Coilpack/compare/1.1.2...HEAD
+
+[1.1.2]: https://github.com/ExpressionEngine/Coilpack/compare/1.1.1...1.1.2
 
 [1.1.1]: https://github.com/ExpressionEngine/Coilpack/compare/1.1.0...1.1.1